go code cleanup
